netomerch/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.8.1
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.9.1
redis==3.5.3
django-redis==5.0.0
django-taggit==1.5.1
django-filter==21.1
drf-yasg==1.20.0
html2text==2020.1.16
celery==5.2.1
django-summernote==0.8.20.0
